+++ punchout_cxml_purchase/models/__init__.py
@@ -3,3 +3,4 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 from . import punchout_session
+from . import purchase_order_line